+++ CHANGELOG.md
@@ -4,7 +4,9 @@
 - Reworked `Lexer`'s code generation
 - Changed `Lexer` constructor `Lexer(dict, input, tokens)` → `Lexer(dict, tokens)`
 - Renamed `Lexer.generate()` method into `Lexer.generateModule()`
-- Added `format` option for `Lexer#generateModule(format)` and `Lexer.generateModule(dict, tokes, format)` methods
+- Removed support for `moduleName` in `Lexer#generateModule()` and `Lexer.generateModule()` methods
+- Added `format` option for `Lexer#generateModule(format)` and `Lexer.generateModule(dict, tokes, format)` methods. Supported formats are `cjs`, `esm` and `iife`
+- Changed `Generator#generateModule(opts)` → `Generator#generateModule(format)`
 
 ## 0.4.18-remastered.1 (2020-11-10)